requirements.txt
accelerate
datasets
einops
elasticsearch
elasticsearch-async
faiss-gpu
fuzzysearch
gensim
gymnasium
imutils
langchain
matplotlib
networkx
notebook==6.5.4
numpy
pandas
pymupdf
pyspellchecker
python-levenshtein
redis
qdrant-client
sentence-transformers
scikit-image
tensorboard
tiktoken
torchmetrics
torchsummary
torchtyping
transformers
unidecode
xformers
